Applied the requested changes.
